move internal types to internal package
